type[] changes to uint32 - based on required change in utype.c and utype.h several routines needed to be changed from uint16 to uint32 to capture all the character flags used in type[]. Also found a few extra bugs which are now corrected here. These corrections will need to be tried by someone who is more familiar with left to right and right to left string quirks just to be sure the fixes given here are good.

+++ inc/gdraw.h
@@ -57,8 +57,8 @@ typedef struct {
 typedef struct gbidata {
     unichar_t *text;
     uint8 *level;
-    uint8 *override;
-    uint16 *type;
+    uint8 *override;	/* override (L2R or R2L) text direction, 1=L2R or -1=R2L */
+    uint32 *type;	/* unicode char feature data, see #define for (uint32)(utype[]) */
     unichar_t **original;
     int32 len;
     unsigned int base_right_to_left: 1;
